refactor: Create internal/siencoding package and move encoding related sources to the package
