+++ api.go
@@ -1,30 +1,1 @@
 package si
-
-import (
-	"bytes"
-	"io"
-
-	"github.com/wonksing/si/v2/internal"
-)
-
-// DecodeJson read src with json bytes then decode it into dst.
-func DecodeJson(dst any, src io.Reader) error {
-	return internal.DecodeJson(dst, src)
-}
-
-// DecodeJsonCopied read src with json bytes then decode it into dst.
-// It also write the data read from src into a bytes.Buffer then returns it.
-func DecodeJsonCopied(dst any, src io.Reader) (*bytes.Buffer, error) {
-	return internal.DecodeJsonCopied(dst, src)
-}
-
-// EncodeJson encode src into json bytes then write to dst.
-func EncodeJson(dst io.Writer, src any) error {
-	return internal.EncodeJson(dst, src)
-}
-
-// EncodeJsonCopied encode src into json bytes then write to dst.
-// It also write encoded bytes of src to a bytes.Buffer then returns it.
-func EncodeJsonCopied(dst io.Writer, src any) (*bytes.Buffer, error) {
-	return internal.EncodeJsonCopied(dst, src)
-}
