+++ sielastic/elastic.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/elastic/go-elasticsearch/v8/esapi"
+	"github.com/wonksing/si/v2/codec"
 	"github.com/wonksing/si/v2/internal"
 )
 
@@ -53,7 +54,7 @@ func (c *Client) IndexDocument(ctx context.Context, indexName string, body []byt
 	}
 
 	var r map[string]interface{}
-	if err := internal.DecodeJson(&r, res.Body); err != nil {
+	if err := codec.DecodeJson(&r, res.Body); err != nil {
 		return nil, err
 	}
 
@@ -117,7 +118,7 @@ func (c *Client) SearchDocuments(ctx context.Context, indexName string, body map
 	buf := internal.GetBytesBuffer(nil)
 	defer internal.PutBytesBuffer(buf)
 
-	if err := internal.EncodeJson(buf, body); err != nil {
+	if err := codec.EncodeJson(buf, body); err != nil {
 		return err
 	}
 
@@ -136,18 +137,18 @@ func (c *Client) SearchDocuments(ctx context.Context, indexName string, body map
 
 	if res.IsError() {
 		resErr := Resp{}
-		copied, err := internal.DecodeJsonCopied(&resErr, res.Body)
+		copied, err := codec.DecodeJsonCopied(&resErr, res.Body)
 		if err != nil {
 			return err
 		}
 
-		if err := internal.DecodeJson(&dest, copied); err != nil {
+		if err := codec.DecodeJson(&dest, copied); err != nil {
 			return err
 		}
 		return errors.New(resErr.Error.Reason)
 	}
 
-	if err := internal.DecodeJson(dest, res.Body); err != nil {
+	if err := codec.DecodeJson(dest, res.Body); err != nil {
 		return err
 	}
